Day 5: Mesh shaders!
Today we've written our first mesh shader that uses meshlet data
prepared on the CPU. We are *not* utilizing the GPU efficiently yet as
our mesh shader is only using one thread on the warp, which should run
at ~3% utilization, and we are *not* optimizing the mesh well which
reflects poorly on the number of meshlets. We will fix this tomorrow.